WIP requires rewrite

+++ reset.sh
@@ -17,21 +17,24 @@ readonly SCRIPT_NAME SCRIPT_DIR
 
 # Accept envronment variable, or fallback to the script's directory
 WMM_HOSTNAME=${WMM_HOSTNAME:-$(basename $PWD)}
-read -p "Purge the current VM $WMM_HOSTNAME? (y/N)" -n 1 -r
-echo    # (optional) move to a new line
-if [[ $REPLY =~ ^[Yy]$ ]]
+
+if multipass list --format csv | grep -q "$WMM_HOSTNAME,"
 then
-	echo "Delete and purge the VM..."
-	multipass delete $WMM_HOSTNAME; multipass purge
+  read -p "Purge the existing VM $WMM_HOSTNAME? (y/N) " -n 1 -r
+  echo    # (optional) move to a new line
+  if [[ $REPLY =~ ^[Yy]$ ]]
+  then
+    echo "Delete and purge the VM..."
+    multipass delete $WMM_HOSTNAME; multipass purge
+  fi
 fi
 
-
-read -p "Replace the scripts and launch? (y/N)" -n 1 -r
+read -p "Replace enabled scripts and launch? (y/N) " -n 1 -r
 echo    # (optional) move to a new line
 if [[ $REPLY =~ ^[Yy]$ ]]
 then
 	# Replace all the scripts
-	rm -f $SCRIPT_DIR/scripts-enabled/*.sh
-	cp $SCRIPT_DIR/scripts-available/*.sh $SCRIPT_DIR/scripts-enabled/
-	$SCRIPT_DIR/host.sh
+	rm -rf scripts-enabled/*
+	cp -r scripts-available/* scripts-enabled/
+  ./host.sh
 fi
